pdf2jpg-service/requirements.txt
fastapi
uvicorn
pymupdf4llm
pillow
python-multipart
psycopg2-binary
sqlalchemy
requests

